fix: conserta envio do formulário de edição de perfil e botão de cancelar
